Partially fix moved/gone db issue.
the idaplugin contains a netnode which saves the project id (pk) and tries to access it
if it's invalid, the server would return 400 we would faecpalm

This temp fix basically tries to see if we got "Invalid pk" from the server and warn the user
that the server may have been changed.
We should handle this better in future releases.